+++ openfisca_france/model/prelevements_obligatoires/prelevements_sociaux/cotisations_sociales/chomage.py
@@ -1,0 +1,32 @@
+from openfisca_france.model.base import Individu, Variable, MONTH, \
+    set_input_divide_by_period, round_, where
+
+
+class chomage_cotisation_retraite_complementaire_journaliere(Variable):
+    value_type = float
+    entity = Individu
+    label = 'Montant de la Cotisation de Retraite Complémentaire (CRC) sur allocation chômage'
+    definition_period = MONTH
+    set_input = set_input_divide_by_period
+    reference = [
+        'https://www.unedic.org/indemnisation/fiches-thematiques/cumul-allocation-salaire',
+        'https://www.unedic.org/indemnisation/fiches-thematiques/retenues-sociales-sur-les-allocations'
+        ]
+
+    def formula(individu, period, parameters):
+        allocation_retour_emploi_journaliere = individu('allocation_retour_emploi_journaliere', period)
+        salaire_journalier_reference_are = individu('salaire_journalier_reference_are', period)
+
+        # le seuil d'exonération de cette cotisation est indexé sur le montant minimum d'ARE
+        seuil_exoneration = parameters(period).chomage.allocation_retour_emploi.montant_minimum_hors_mayotte
+        
+        taux_cotisation = parameters(period).chomage.complement_are.taux_crc
+        cotisation_theorique = salaire_journalier_reference_are * taux_cotisation
+        allocation_cotisation_deduite = allocation_retour_emploi_journaliere - cotisation_theorique
+
+        return round_(
+            where(
+                allocation_cotisation_deduite > seuil_exoneration,
+                -1 * cotisation_theorique,
+                0),
+            2)
